Update: auto clear trash.

+++ app/src/main/java/ink/snowland/wkuwku/ui/game/GamesViewModel.java
@@ -49,10 +49,10 @@ public class GamesViewModel extends BaseViewModel {
         return mAllGames;
     }
 
-    private void copyFiles(@NonNull String filename, @NonNull Uri uri, @NonNull ActionListener listener) {
+    private void copyFiles(@NonNull Uri uri, @NonNull File file, @NonNull ActionListener listener) {
         if ("https".equals(uri.getScheme())) {
             setPendingIndicator(true, R.string.please_wait);
-            DownloadManager.newRequest(uri.toString(), FileManager.getFile(FileManager.ROM_DIRECTORY, filename))
+            DownloadManager.newRequest(uri.toString(), file)
                     .doOnProgressUpdate((progress, max) -> {
                         setPendingMessage(getString(R.string.fmt_downloading, (float) progress / max * 100));
                     })
@@ -61,7 +61,7 @@ public class GamesViewModel extends BaseViewModel {
                             setPendingMessage(R.string.connecting);
                         }
                     })
-                    .doOnComplete(file -> {
+                    .doOnComplete(it -> {
                         listener.onSuccess();
                     })
                     .doOnFinally(() -> {
@@ -72,12 +72,10 @@ public class GamesViewModel extends BaseViewModel {
         } else {
             Completable.create(emitter -> {
                         setPendingIndicator(true, R.string.copying_files);
-                        File to = FileManager.getFile(FileManager.ROM_DIRECTORY, filename);
-                        boolean noError;
+                        boolean noError = false;
                         try (InputStream from = getApplication().getContentResolver().openInputStream(uri)) {
-                            noError = from != null;
                             if (from != null) {
-                                noError = FileUtils.copy(from, to);
+                                noError = FileUtils.copy(from, file);
                             }
                         }
                         if (noError) {
@@ -97,68 +95,56 @@ public class GamesViewModel extends BaseViewModel {
     }
 
     public void addGame(@NonNull Game game, @NonNull Uri uri) {
-        final String filename = game.filepath;
-        int infoMask = ArchiveUtils.getFileInfoMask(filename);
-        if ((infoMask & ArchiveUtils.FLAG_ARCHIVE_FILE_TYPE) == ArchiveUtils.FLAG_ARCHIVE_FILE_TYPE
-                && (infoMask & ArchiveUtils.FLAG_SUPPORTED_ARCHIVE_FILE_TYPE) != ArchiveUtils.FLAG_SUPPORTED_ARCHIVE_FILE_TYPE) {
-            post(() -> {
-                Toast.makeText(getApplication(), R.string.unsupported_archive_format, Toast.LENGTH_SHORT).show();
-            });
+        final File file = getFile(ROM_DIRECTORY, game.filepath);
+        boolean isArchiveType = ArchiveUtils.isArchiveType(file);
+        if (isArchiveType && !ArchiveUtils.isSupported(file)) {
+            Toast.makeText(getApplication(), R.string.unsupported_archive_format, Toast.LENGTH_SHORT).show();
             return;
         }
-        copyFiles(filename, uri, new ActionListener() {
+        copyFiles(uri, file, new ActionListener() {
             @Override
             public void onSuccess() {
-                boolean noError = true;
                 IEmulator emulator = EmulatorManager.getDefaultEmulator(game.system);
-                File file = FileManager.getFile(FileManager.ROM_DIRECTORY, game.filepath);
-                if (emulator == null) return;
-                if ((infoMask & ArchiveUtils.FLAG_ARCHIVE_FILE_TYPE) == ArchiveUtils.FLAG_ARCHIVE_FILE_TYPE
-                        && (infoMask & ArchiveUtils.FLAG_SUPPORTED_ARCHIVE_FILE_TYPE) == ArchiveUtils.FLAG_SUPPORTED_ARCHIVE_FILE_TYPE) {
-                    boolean emulatorSupportedArchive = emulator.searchSupportedContent(file) != null;
-                    if (!emulatorSupportedArchive) {
-                        setPendingIndicator(true, R.string.unzipping_files);
-                        File originFile = file;
-                        try {
-                            String unzippedPath = ArchiveUtils.extract(originFile);
-                            file = new File(unzippedPath);
-                        } catch (IOException e) {
-                            e.printStackTrace(System.err);
-                            if (e instanceof FileAlreadyExistsException) {
-                                post(() -> {
-                                    Toast.makeText(getApplication(), getString(R.string.file_already_exists), Toast.LENGTH_SHORT).show();
-                                });
+                File content = emulator.searchSupportedContent(file);
+                if (content != null) {
+                    game.filepath = content.getAbsolutePath();
+                    game.addedTime = System.currentTimeMillis();
+                    game.lastModifiedTime = game.addedTime;
+                    game.state = Game.STATE_VALID;
+                    game.md5 = FileUtils.getMD5Sum(file);
+                    insert(game);
+                } else if (isArchiveType) {
+                    setPendingIndicator(true, R.string.unzipping_files);
+                    File outdir = new File(file.getParentFile(), FileUtils.getNameNotExtension(file));
+                    ArchiveUtils.asyncExtract(file, outdir, new ActionListener() {
+                        @Override
+                        public void onSuccess() {
+                            File content = emulator.searchSupportedContent(outdir);
+                            if (content != null) {
+                                game.filepath = content.getAbsolutePath();
+                                game.addedTime = System.currentTimeMillis();
+                                game.lastModifiedTime = game.addedTime;
+                                game.state = Game.STATE_VALID;
+                                game.md5 = FileUtils.getMD5Sum(file);
+                                insert(game);
                             } else {
-                                post(() -> {
-                                    showErrorToast(e);
-                                });
+                                FileUtils.delete(outdir);
+                                Toast.makeText(getApplication(), R.string.could_not_find_valid_rom_file, Toast.LENGTH_LONG).show();
                             }
-                            noError = false;
-                        } finally {
-                            FileUtils.delete(originFile);
+                            FileUtils.delete(file);
+                            setPendingIndicator(false);
                         }
-                    }
-                }
-                if (!noError) return;
-                if (file.isDirectory()) {
-                    File romFile = emulator.searchSupportedContent(file);
-                    if (romFile == null) {
-                        FileUtils.delete(file);
-                    }
-                    file = romFile;
-                }
-                if (file == null || !file.exists()) {
-                    post(() -> {
-                        Toast.makeText(getApplication(), R.string.could_not_find_valid_rom_file, Toast.LENGTH_SHORT).show();
+
+                        @Override
+                        public void onFailure(Throwable e) {
+                            FileUtils.delete(file);
+                            setPendingIndicator(false);
+                            Toast.makeText(getApplication(), R.string.failed_extract, Toast.LENGTH_LONG).show();
+                        }
                     });
-                    return;
+                } else {
+                    Toast.makeText(getApplication(), R.string.could_not_find_valid_rom_file, Toast.LENGTH_LONG).show();
                 }
-                game.filepath = file.getAbsolutePath();
-                game.addedTime = System.currentTimeMillis();
-                game.lastModifiedTime = game.addedTime;
-                game.state = Game.STATE_VALID;
-                game.md5 = FileUtils.getMD5Sum(file);
-                insert(game);
             }
 
             @Override
@@ -169,53 +155,44 @@ public class GamesViewModel extends BaseViewModel {
     }
 
     public void update(@NonNull Game game) {
-        Disposable disposable = AppDatabase.db.gameInfoDao().update(game)
+        AppDatabase.db.gameInfoDao().update(game)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .doOnError(error -> {
-                    showErrorToast(error);
-                    error.printStackTrace(System.err);
-                })
-                .subscribe(() -> {
-                }, error -> {/*Ignored*/});
+                .doOnError(this::onError)
+                .onErrorComplete()
+                .subscribe();
     }
 
     public void delete(@NonNull Game game) {
-        Disposable disposable = AppDatabase.db.gameInfoDao().delete(game)
+        AppDatabase.db.gameInfoDao().delete(game)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .doOnError(error -> {
-                    showErrorToast(error);
-                    error.printStackTrace(System.err);
-                })
-                .subscribe(() -> {
-                    FileUtils.delete(game.filepath);
-                }, error -> {/*Ignored*/});
+                .doOnComplete(() -> FileUtils.delete(game.filepath))
+                .doOnError(this::onError)
+                .onErrorComplete()
+                .subscribe();
     }
 
     private void insert(@NonNull Game game) {
-        Disposable disposable = AppDatabase.db.gameInfoDao()
+        AppDatabase.db.gameInfoDao()
                 .insert(game)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .doOnError(error -> {
                     if (!(error instanceof SQLiteConstraintException)) {
                         File parent = new File(game.filepath).getParentFile();
-                        assert parent != null;
-                        if (!parent.equals(getFileDirectory(ROM_DIRECTORY))) {
+                        if (parent != null && !parent.equals(getFileDirectory(ROM_DIRECTORY))) {
                             FileUtils.delete(parent);
                         } else {
                             FileUtils.delete(game.filepath);
                         }
-                        post(() -> {
-                            Toast.makeText(getApplication(), getString(R.string.fmt_game_already_exists, game.title), Toast.LENGTH_SHORT).show();
-                        });
+                        Toast.makeText(getApplication(), getString(R.string.fmt_game_already_exists, game.title), Toast.LENGTH_SHORT).show();
                     } else {
-                        showErrorToast(error);
+                        this.onError(error);
                     }
                 })
-                .subscribe(() -> {
-                }, error -> {/*Ignored*/});
+                .onErrorComplete()
+                .subscribe();
     }
 
     @Override
